requirements.txt
nltk==3.4
numpy==1.16.0
pandas==0.24.0
python-dateutil==2.7.5
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
tweet-preprocessor==0.5.0
